Safari-611.1.10/LayoutTests/js/dfg-use-function-as-variable-with-closure-expected.txt

Tests what happens when you use a function as a variable, and that function is closed over.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS successfullyParsed is true

TEST COMPLETE
